+++ mayan/settings/base.py
@@ -106,6 +106,7 @@ INSTALLED_APPS = (
     'mayan.apps.permissions.apps.PermissionsApp',
     'mayan.apps.platform.apps.PlatformApp',
     'mayan.apps.platform_docker.apps.PlatformDockerApp',
+    'mayan.apps.platform_gitlab.apps.PlatformGitlabApp',
     'mayan.apps.quotas.apps.QuotasApp',
     'mayan.apps.rest_api.apps.RESTAPIApp',
     'mayan.apps.smart_settings.apps.SmartSettingsApp',